refactor(model): move prompt envelopes into core and enforce import contracts
Third task of the modular-monolith work, and the one that makes the
boundary rule checkable rather than aspirational.

PromptEnvelope and ResponseEnvelope described a channel-neutral payload
but lived in primer.channel.adapter, so primer.agent.inform and
primer.worker.yield_runtime imported the channel package to name a type
that has nothing to do with channels. They move to primer.model.envelope
verbatim; primer.channel.adapter re-exports both, because a dozen channel
adapters and their tests import them from there and the move is not
supposed to be a rename.

m001's knowledge import becomes lazy for the same reason: storage is core
and knowledge is a subsystem, so a module-scope import made every install
pull the knowledge stack just to register the migration chain.

The contracts are the point. Two forbidden-import contracts, checked by
import-linter in CI right after ruff, over 523 files and 2273
dependencies: core layers (model/int/storage/common/claim) may not import
the optional subsystems, and agent/worker may not import channel. Both
police DIRECT imports; indirect chains through the app layer are expected.

Three sanctioned seams are listed with reasons rather than the contracts
being widened to accommodate them, so the ignore list stays an audit
trail of deliberate exceptions instead of a way to make the check pass.
